% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_hist_function.R
\name{grade_hist}
\alias{grade_hist}
\title{grade_hist}
\usage{
grade_hist(df, grade_col)
}
\arguments{
\item{df}{Dataframe with a column with grades}

\item{grade_col}{Column name that contains the grades}
}
\value{
plot
}
\description{
creates histogram plot for grades
}
